+++ util/analysis/helpers/printOffEdges.m
@@ -1,0 +1,5 @@
+function IDs = printOffEdges(neuron)
+    IDs = neuron.offEdges;
+    
+    IDs(ismember(IDs, neuron.terminals)) = [];
+