+++ util/analysis/iplDepth.m
@@ -20,12 +20,17 @@ function [iplPercent, stats] = iplDepth(Neuron, INL, GCL, numBins)
 	% History
 	%	7Feb2018 - SSP
     %   19Feb2018 - SSP - Added numBins input
+    %   22Oct2018 - SSP - Added boundary markers from cache
 	% ---------------------------------------------------------------------
 
 	assert(isa(Neuron, 'sbfsem.core.StructureAPI'),...
 		'Input a StructureAPI object');
     if nargin < 4
         numBins = 20;
+    end
+    if nargin < 3
+        GCL = sbfsem.builtin.GCLBoundary(Neuron.source, true);
+        INL = sbfsem.builtin.INLBoundary(Neuron.source, true);
     end
 
 	nodes = Neuron.getCellNodes;
@@ -42,7 +47,8 @@ function [iplPercent, stats] = iplDepth(Neuron, INL, GCL, numBins)
 	vINL = interp2(X, Y, INL.interpolatedSurface,...
 		xyz(:,1), xyz(:,2));
 
-	iplPercent = (xyz(:,3) - vGCL)./((vINL - vGCL)+eps);
+	% iplPercent = (xyz(:,3) - vGCL)./((vINL - vGCL)+eps);
+    iplPercent = (xyz(:, 3) - vINL) ./ ((vGCL - vINL)+eps);
 	iplPercent(isnan(iplPercent)) = [];
     disp('Mean +- SEM microns (n):');
 	printStat(iplPercent');
@@ -54,12 +60,21 @@ function [iplPercent, stats] = iplDepth(Neuron, INL, GCL, numBins)
 	stats.n = numel(iplPercent);
 	fprintf('Median IPL Depth = %.3g\n', stats.median);
 
-	figure();
-	hist(iplPercent, numBins);
-	xlim([-1, 1]);
-    title(sprintf('IPL depth estimates for c%u', Neuron.ID));
-	ylabel('Number of annotations');
-	xlabel('Percent IPL Depth');
+	ax = axes('Parent', figure());
+	hist(ax, iplPercent, numBins);
+
+    x = get(ax, 'XLim');
+    if x(1) > 0
+        x(1) = 0;
+    end
+    if x(2) < 1
+        x(2) = 1;
+    end
+    xlim(ax, x);
+    
+    title(ax, sprintf('IPL depth estimates for c%u', Neuron.ID));
+	ylabel(ax, 'Number of annotations');
+	xlabel(ax, 'Percent IPL Depth');
 
 	figure();
 	hist(vINL-vGCL, numBins); hold on;